pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "city-rides-lakehouse"
version = "0.1.0"
description = "End-to-end data engineering project using Python, DuckDB, dbt, Streamlit, and Prefect."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "pyarrow>=14",
  "duckdb>=0.10",
  "python-dateutil>=2.9",
  "python-dotenv>=1.0",
  "prefect>=2.14",
  "pandera>=0.18",
  "streamlit>=1.30",
  "altair>=5.0",
  "dbt-core>=1.7",
  "dbt-duckdb>=1.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
city-rides-pipeline = "src.pipeline.run_pipeline:main"
city-rides-flow = "src.pipeline.prefect_flow:main"

[tool.setuptools.packages.find]
include = ["src*"]
